requirements.txt

python-telegram-bot==20.3
requests==2.31.0
python-dotenv==1.0.1
beautifulsoup4==4.12.2
lxml==4.9.3
html5lib==1.1
markdown==3.4.3
markdownify==0.14.0
nltk==3.8.1
tiktoken==0.4.0
openai==0.27.8